+++ lib/src/enums.dart
@@ -2,6 +2,13 @@
 /// Each type has a unique color and icon.
 enum NotificationType { success, error, warning, info }
 
-enum NotificationPosition { topRight, topLeft, bottomRight, bottomLeft }
+enum NotificationPosition {
+  topRight,
+  topLeft,
+  topCenter,
+  bottomRight,
+  bottomLeft,
+  bottomCenter
+}
 
 enum SlideDirection { leftToRight, rightToLeft, topToBottom, bottomToTop }